@@ -2,7 +2,7 @@ package com.rapsealk.agroundcontrol
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
 import com.google.android.gms.location.FusedLocationProviderClient
